Workaround for windows/github/zip: fix linked ./tests in ./inst/
./inst/tests/ was a symlink to ./tests/. On windows, after installation via
install_github(), we I an error from file.copy(), which can't find tests.

Unzipping the archive downloaded from github on windows, the link is broken.
Unzipping the archive downloaded from github on *nix, the link works.
Unzipping the tarball built by R CMD  on windows, the link is broken.
Unzipping the tarball built by R CMD  on *nix, the link works.

So it's a windows/github/zip issue.
I hope to fix it by moving the tests to be installed into inst/ and linking
them, so move ./tests/ to ./inst/ and link it here.

- DESCRIPTION: set Version to 0.1-2
- configure: rename make target purge to clean_git (as it git cleans), add make
  target purge as clean, remove and clean_git.