add a "Depends:"-field to the temporary package's DESCRIPTION file
If a function used a a function from some package, the documentation of the
temporary package failed (for example, if the function's arguments where checked
with functions from the checkmate package). So I add the option add a
"Depends"-field to its DESCRIPTION file to load the package containing the
missing functions.

- R/package_utils.r: add add_dependencies_to_description()
- DESCRIPTION: bump date and version.
- demo/documentation.r: fix the demo
- R/create_roxygen_documentation.r (create_roxygen_documentation): call
  add_dependencies_to_description, add dependencies argument for
  add_dependencies_to_description and a checkmate check for that argument.
- man/*: new docs.